Time: N/A (0%), Space: N/A (0%) - LeetHub

+++ 1482-minimum-number-of-days-to-make-m-bouquets/1482-minimum-number-of-days-to-make-m-bouquets.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int minDays(vector<int>& bloomDay, int m, int k) {
-        if(m*k>bloomDay.size()){
+        if((long long)m*k>bloomDay.size()){
             return -1;
         }
         int low=1;
